Update various cardinality and length constraints to use unsigned integer types

+++ include/valijson/validation_visitor.hpp
@@ -422,8 +422,8 @@ public:
         }
 
         const std::string s = target.asString();
-        const int64_t len = utils::u8_strlen(s.c_str());
-        const int64_t maxLength = constraint.getMaxLength();
+        const uint64_t len = utils::u8_strlen(s.c_str());
+        const uint64_t maxLength = constraint.getMaxLength();
         if (len <= maxLength) {
             return true;
         }
@@ -451,7 +451,7 @@ public:
             return true;
         }
 
-        const int64_t maxProperties = constraint.getMaxProperties();
+        const uint64_t maxProperties = constraint.getMaxProperties();
 
         if (target.asObject().size() <= maxProperties) {
             return true;
@@ -518,7 +518,7 @@ public:
             return true;
         }
 
-        const int64_t minItems = constraint.getMinItems();
+        const uint64_t minItems = constraint.getMinItems();
         if (target.asArray().size() >= minItems) {
             return true;
         }
@@ -545,8 +545,8 @@ public:
         }
 
         const std::string s = target.asString();
-        const int len = utils::u8_strlen(s.c_str());
-        const int64_t minLength = constraint.getMinLength();
+        const uint64_t len = utils::u8_strlen(s.c_str());
+        const uint64_t minLength = constraint.getMinLength();
         if (len >= minLength) {
             return true;
         }
@@ -574,7 +574,7 @@ public:
             return true;
         }
 
-        const int64_t minProperties = constraint.getMinProperties();
+        const uint64_t minProperties = constraint.getMinProperties();
 
         if (target.asObject().size() >= minProperties) {
             return true;
